multiple updates
Improve csv write time, rename files
- separate csv for each month
- optimize count_GLM_min
- more efficient checks for missing data
- switch from xarray to netCDF4
- combine time columns to single timestamp
