+++ radar_locations.py
@@ -1,0 +1,1325 @@
+radars = {
+    "KABR": {
+        "wban": "14929",
+        "staid": "KABR",
+        "stname": "Aberdeen, SD",
+        "latitude": "45.45583333",
+        "longitude": "-98.41305556",
+        "elevation": "1302",
+        "height": "20"
+    },
+    "KENX": {
+        "wban": "54766",
+        "staid": "KENX",
+        "stname": "Albany, NY",
+        "latitude": "42.58638889",
+        "longitude": "-74.06388889",
+        "elevation": "1826",
+        "height": "20"
+    },
+    "KABX": {
+        "wban": "3019",
+        "staid": "KABX",
+        "stname": "Albuquerque, NM",
+        "latitude": "35.14972222",
+        "longitude": "-106.8238889",
+        "elevation": "5870",
+        "height": "20"
+    },
+    "KFDR": {
+        "wban": "3981",
+        "staid": "KFDR",
+        "stname": "Altus AFB, OK",
+        "latitude": "34.36222222",
+        "longitude": "-98.97638889",
+        "elevation": "1267",
+        "height": "10"
+    },
+    "KAMA": {
+        "wban": "23047",
+        "staid": "KAMA",
+        "stname": "Amarillo, TX",
+        "latitude": "35.23333333",
+        "longitude": "-101.7091667",
+        "elevation": "3587",
+        "height": "20"
+    },
+    "KFFC": {
+        "wban": "53819",
+        "staid": "KFFC",
+        "stname": "Atlanta, GA",
+        "latitude": "33.36361111",
+        "longitude": "-84.56583333",
+        "elevation": "858",
+        "height": "30"
+    },
+    "KEWX": {
+        "wban": "12971",
+        "staid": "KEWX",
+        "stname": "Austin/San Antonio, TX",
+        "latitude": "29.70388889",
+        "longitude": "-98.02833333",
+        "elevation": "633",
+        "height": "20"
+    },
+    "KBBX": {
+        "wban": "93240",
+        "staid": "KBBX",
+        "stname": "Beale AFB, CA",
+        "latitude": "39.49611111",
+        "longitude": "-121.6316667",
+        "elevation": "167",
+        "height": "10"
+    },
+    "KBLX": {
+        "wban": "94096",
+        "staid": "KBLX",
+        "stname": "Billings, MT",
+        "latitude": "45.85388889",
+        "longitude": "-108.6066667",
+        "elevation": "3598",
+        "height": "10"
+    },
+    "KBGM": {
+        "wban": "4725",
+        "staid": "KBGM",
+        "stname": "Binghamton, NY",
+        "latitude": "42.19972222",
+        "longitude": "-75.98472222",
+        "elevation": "1606",
+        "height": "20"
+    },
+    "KBMX": {
+        "wban": "53823",
+        "staid": "KBMX",
+        "stname": "Birmingham, AL",
+        "latitude": "33.17222222",
+        "longitude": "-86.76972222",
+        "elevation": "645",
+        "height": "30"
+    },
+    "KBIS": {
+        "wban": "24011",
+        "staid": "KBIS",
+        "stname": "Bismarck, ND",
+        "latitude": "46.77083333",
+        "longitude": "-100.7605556",
+        "elevation": "1658",
+        "height": "20"
+    },
+    "KCBX": {
+        "wban": "4101",
+        "staid": "KCBX",
+        "stname": "Boise, ID",
+        "latitude": "43.49055556",
+        "longitude": "-116.2355556",
+        "elevation": "3061",
+        "height": "20"
+    },
+    "KBOX": {
+        "wban": "54765",
+        "staid": "KBOX",
+        "stname": "Boston, MA",
+        "latitude": "41.95583333",
+        "longitude": "-71.13694444",
+        "elevation": "118",
+        "height": "30"
+    },
+    "KBRO": {
+        "wban": "12919",
+        "staid": "KBRO",
+        "stname": "Brownsville, TX",
+        "latitude": "25.91611111",
+        "longitude": "-97.41888889",
+        "elevation": "23",
+        "height": "10"
+    },
+    "KBUF": {
+        "wban": "14733",
+        "staid": "KBUF",
+        "stname": "Buffalo, NY",
+        "latitude": "42.94888889",
+        "longitude": "-78.73666667",
+        "elevation": "693",
+        "height": "20"
+    },
+    "KCXX": {
+        "wban": "54774",
+        "staid": "KCXX",
+        "stname": "Burlington, VT",
+        "latitude": "44.51111111",
+        "longitude": "-73.16694444",
+        "elevation": "317",
+        "height": "30"
+    },
+    "KFDX": {
+        "wban": "3022",
+        "staid": "KFDX",
+        "stname": "Cannon AFB, NM",
+        "latitude": "34.63527778",
+        "longitude": "-103.63",
+        "elevation": "4650",
+        "height": "10"
+    },
+    "KICX": {
+        "wban": "PENDING",
+        "staid": "KICX",
+        "stname": "Cedar City, UT",
+        "latitude": "37.59083333",
+        "longitude": "-112.8622222",
+        "elevation": "10600",
+        "height": "30"
+    },
+    "KCLX": {
+        "wban": "53845",
+        "staid": "KCLX",
+        "stname": "Charleston, SC",
+        "latitude": "32.65555556",
+        "longitude": "-81.04222222",
+        "elevation": "97",
+        "height": "30"
+    },
+    "KRLX": {
+        "wban": "53834",
+        "staid": "KRLX",
+        "stname": "Charleston, WV",
+        "latitude": "38.31111111",
+        "longitude": "-81.72305556",
+        "elevation": "1080",
+        "height": "30"
+    },
+    "KCYS": {
+        "wban": "24018",
+        "staid": "KCYS",
+        "stname": "Cheyenne, WY",
+        "latitude": "41.15194444",
+        "longitude": "-104.8061111",
+        "elevation": "6128",
+        "height": "10"
+    },
+    "KLOT": {
+        "wban": "4831",
+        "staid": "KLOT",
+        "stname": "Chicago, IL",
+        "latitude": "41.60472222",
+        "longitude": "-88.08472222",
+        "elevation": "663",
+        "height": "20"
+    },
+    "KILN": {
+        "wban": "13841",
+        "staid": "KILN",
+        "stname": "Cincinnati, OH",
+        "latitude": "39.42027778",
+        "longitude": "-83.82166667",
+        "elevation": "1056",
+        "height": "30"
+    },
+    "KCLE": {
+        "wban": "14820",
+        "staid": "KCLE",
+        "stname": "Cleveland, OH",
+        "latitude": "41.41305556",
+        "longitude": "-81.85972222",
+        "elevation": "763",
+        "height": "20"
+    },
+    "KCAE": {
+        "wban": "13883",
+        "staid": "KCAE",
+        "stname": "Columbia, SC",
+        "latitude": "33.94861111",
+        "longitude": "-81.11833333",
+        "elevation": "231",
+        "height": "30"
+    },
+    "KGWX": {
+        "wban": "53837",
+        "staid": "KGWX",
+        "stname": "Columbus AFB, MS",
+        "latitude": "33.89666667",
+        "longitude": "-88.32888889",
+        "elevation": "476",
+        "height": "20"
+    },
+    "KCRP": {
+        "wban": "12924",
+        "staid": "KCRP",
+        "stname": "Corpus Christi, TX",
+        "latitude": "27.78416667",
+        "longitude": "-97.51111111",
+        "elevation": "45",
+        "height": "20"
+    },
+    "KFWS": {
+        "wban": "3985",
+        "staid": "KFWS",
+        "stname": "Dallas/Ft. Worth, TX",
+        "latitude": "32.57305556",
+        "longitude": "-97.30305556",
+        "elevation": "683",
+        "height": "20"
+    },
+    "KDVN": {
+        "wban": "94982",
+        "staid": "KDVN",
+        "stname": "Davenport, IA",
+        "latitude": "41.61166667",
+        "longitude": "-90.58083333",
+        "elevation": "754",
+        "height": "20"
+    },
+    "KFTG": {
+        "wban": "3018",
+        "staid": "KFTG",
+        "stname": "Denver, CO",
+        "latitude": "39.78666667",
+        "longitude": "-104.5458333",
+        "elevation": "5497",
+        "height": "30"
+    },
+    "KDMX": {
+        "wban": "94984",
+        "staid": "KDMX",
+        "stname": "Des Moines, IA",
+        "latitude": "41.73111111",
+        "longitude": "-93.72277778",
+        "elevation": "981",
+        "height": "30"
+    },
+    "KDTX": {
+        "wban": "4830",
+        "staid": "KDTX",
+        "stname": "Detroit, MI",
+        "latitude": "42.69972222",
+        "longitude": "-83.47166667",
+        "elevation": "1072",
+        "height": "30"
+    },
+    "KDDC": {
+        "wban": "13985",
+        "staid": "KDDC",
+        "stname": "Dodge City, KS",
+        "latitude": "37.76083333",
+        "longitude": "-99.96888889",
+        "elevation": "2590",
+        "height": "20"
+    },
+    "KDOX": {
+        "wban": "93770",
+        "staid": "KDOX",
+        "stname": "Dover AFB, DE",
+        "latitude": "38.82555556",
+        "longitude": "-75.43972222",
+        "elevation": "50",
+        "height": "30"
+    },
+    "KDLH": {
+        "wban": "14913",
+        "staid": "KDLH",
+        "stname": "Duluth, MN",
+        "latitude": "46.83694444",
+        "longitude": "-92.20972222",
+        "elevation": "1428",
+        "height": "30"
+    },
+    "KDYX": {
+        "wban": "3987",
+        "staid": "KDYX",
+        "stname": "Dyess AFB, TX",
+        "latitude": "32.53833333",
+        "longitude": "-99.25444444",
+        "elevation": "1517",
+        "height": "10"
+    },
+    "KEYX": {
+        "wban": "53114",
+        "staid": "KEYX",
+        "stname": "Edwards AFB, CA",
+        "latitude": "35.09777778",
+        "longitude": "-117.5608333",
+        "elevation": "2757",
+        "height": "20"
+    },
+    "KEVX": {
+        "wban": "53825",
+        "staid": "KEVX",
+        "stname": "Eglin AFB, FL",
+        "latitude": "30.56444444",
+        "longitude": "-85.92138889",
+        "elevation": "140",
+        "height": "20"
+    },
+    "KEPZ": {
+        "wban": "3020",
+        "staid": "KEPZ",
+        "stname": "El Paso, TX",
+        "latitude": "31.87305556",
+        "longitude": "-106.6980556",
+        "elevation": "4104",
+        "height": "30"
+    },
+    "KLRX": {
+        "wban": "4108",
+        "staid": "KLRX",
+        "stname": "Elko, NV",
+        "latitude": "40.73972222",
+        "longitude": "-116.8027778",
+        "elevation": "6744",
+        "height": "30"
+    },
+    "KBHX": {
+        "wban": "94289",
+        "staid": "KBHX",
+        "stname": "Eureka, CA",
+        "latitude": "40.49833333",
+        "longitude": "-124.2919444",
+        "elevation": "2402",
+        "height": "30"
+    },
+    "KFSX": {
+        "wban": "53113",
+        "staid": "KFSX",
+        "stname": "Flagstaff, AZ",
+        "latitude": "34.57444444",
+        "longitude": "-111.1977778",
+        "elevation": "7417",
+        "height": "20"
+    },
+    "KHPX": {
+        "wban": "53839",
+        "staid": "KHPX",
+        "stname": "Fort Campbell, KY",
+        "latitude": "36.73666667",
+        "longitude": "-87.285",
+        "elevation": "576",
+        "height": "10"
+    },
+    "KGRK": {
+        "wban": "3992",
+        "staid": "KGRK",
+        "stname": "Fort Hood, TX",
+        "latitude": "30.72194444",
+        "longitude": "-97.38305556",
+        "elevation": "538",
+        "height": "10"
+    },
+    "KPOE": {
+        "wban": "3993",
+        "staid": "KPOE",
+        "stname": "Fort Polk, LA",
+        "latitude": "31.15555556",
+        "longitude": "-92.97583333",
+        "elevation": "408",
+        "height": "10"
+    },
+    "KEOX": {
+        "wban": "53851",
+        "staid": "KEOX",
+        "stname": "Fort Rucker, AL",
+        "latitude": "31.46055556",
+        "longitude": "-85.45944444",
+        "elevation": "434",
+        "height": "10"
+    },
+    "KSRX": {
+        "wban": "PENDING",
+        "staid": "KSRX",
+        "stname": "Fort Smith, AR",
+        "latitude": "35.29055556",
+        "longitude": "-94.36166667",
+        "elevation": "640",
+        "height": "20"
+    },
+    "KIWX": {
+        "wban": "PENDING",
+        "staid": "KIWX",
+        "stname": "Fort Wayne, IN",
+        "latitude": "41.35888889",
+        "longitude": "-85.7",
+        "elevation": "960",
+        "height": "25"
+    },
+    "KAPX": {
+        "wban": "4837",
+        "staid": "KAPX",
+        "stname": "Gaylord, MI",
+        "latitude": "44.90722222",
+        "longitude": "-84.71972222",
+        "elevation": "1464",
+        "height": "20"
+    },
+    "KGGW": {
+        "wban": "94008",
+        "staid": "KGGW",
+        "stname": "Glasgow, MT",
+        "latitude": "48.20638889",
+        "longitude": "-106.625",
+        "elevation": "2276",
+        "height": "20"
+    },
+    "KGLD": {
+        "wban": "23065",
+        "staid": "KGLD",
+        "stname": "Goodland, KS",
+        "latitude": "39.36694444",
+        "longitude": "-101.7002778",
+        "elevation": "3651",
+        "height": "10"
+    },
+    "KMVX": {
+        "wban": "94986",
+        "staid": "KMVX",
+        "stname": "Grand Forks, ND",
+        "latitude": "47.52777778",
+        "longitude": "-97.32555556",
+        "elevation": "986",
+        "height": "20"
+    },
+    "KGJX": {
+        "wban": "3025",
+        "staid": "KGJX",
+        "stname": "Grand Junction, Co",
+        "latitude": "39.06222222",
+        "longitude": "-108.2138889",
+        "elevation": "9991",
+        "height": "10"
+    },
+    "KGRR": {
+        "wban": "94860",
+        "staid": "KGRR",
+        "stname": "Grand Rapids, MI",
+        "latitude": "42.89388889",
+        "longitude": "-85.54472222",
+        "elevation": "778",
+        "height": "20"
+    },
+    "KTFX": {
+        "wban": "4102",
+        "staid": "KTFX",
+        "stname": "Great Falls, MT",
+        "latitude": "47.45972222",
+        "longitude": "-111.3852778",
+        "elevation": "3714",
+        "height": "10"
+    },
+    "KGRB": {
+        "wban": "14898",
+        "staid": "KGRB",
+        "stname": "Green Bay, WI",
+        "latitude": "44.49833333",
+        "longitude": "-88.11138889",
+        "elevation": "682",
+        "height": "30"
+    },
+    "KGSP": {
+        "wban": "3870",
+        "staid": "KGSP",
+        "stname": "Greer, SC",
+        "latitude": "34.88333333",
+        "longitude": "-82.22",
+        "elevation": "940",
+        "height": "30"
+    },
+    "KRMX": {
+        "wban": "54763",
+        "staid": "KRMX",
+        "stname": "Griffiss AFB, NY",
+        "latitude": "43.46777778",
+        "longitude": "-75.45777778",
+        "elevation": "1516",
+        "height": "30"
+    },
+    "KUEX": {
+        "wban": "94981",
+        "staid": "KUEX",
+        "stname": "Hastings, NE",
+        "latitude": "40.32083333",
+        "longitude": "-98.44194444",
+        "elevation": "1976",
+        "height": "20"
+    },
+    "KHDX": {
+        "wban": "3023",
+        "staid": "KHDX",
+        "stname": "Holloman AFB, NM",
+        "latitude": "33.07638889",
+        "longitude": "-106.1227778",
+        "elevation": "4222",
+        "height": "10"
+    },
+    "KCBW": {
+        "wban": "94625",
+        "staid": "KCBW",
+        "stname": "Houlton, ME",
+        "latitude": "46.03916667",
+        "longitude": "-67.80638889",
+        "elevation": "746",
+        "height": "30"
+    },
+    "KHGX": {
+        "wban": "3980",
+        "staid": "KHGX",
+        "stname": "Houston/Galveston, TX",
+        "latitude": "29.47194444",
+        "longitude": "-95.07916667",
+        "elevation": "18",
+        "height": "20"
+    },
+    "KHTX": {
+        "wban": "PENDING",
+        "staid": "KHTX",
+        "stname": "Huntsville, AL",
+        "latitude": "34.93055556",
+        "longitude": "-86.08333333",
+        "elevation": "1760",
+        "height": "15"
+    },
+    "KIND": {
+        "wban": "93819",
+        "staid": "KIND",
+        "stname": "Indianapolis, IN",
+        "latitude": "39.7075",
+        "longitude": "-86.28027778",
+        "elevation": "790",
+        "height": "20"
+    },
+    "KJKL": {
+        "wban": "3889",
+        "staid": "KJKL",
+        "stname": "Jackson, KY",
+        "latitude": "37.59083333",
+        "longitude": "-83.31305556",
+        "elevation": "1364",
+        "height": "25"
+    },
+    "KJAN": {
+        "wban": "3940",
+        "staid": "KJAN",
+        "stname": "Jackson, MS",
+        "latitude": "32.31777778",
+        "longitude": "-90.08",
+        "elevation": "297",
+        "height": "20"
+    },
+    "KJAX": {
+        "wban": "13889",
+        "staid": "KJAX",
+        "stname": "Jacksonville, FL",
+        "latitude": "30.48472222",
+        "longitude": "-81.70194444",
+        "elevation": "33",
+        "height": "20"
+    },
+    "PHKN": {
+        "wban": "PENDING",
+        "staid": "PHKN",
+        "stname": "Kamuela, HI",
+        "latitude": "20.12555556",
+        "longitude": "-155.7777778",
+        "elevation": "3812",
+        "height": "30"
+    },
+    "KEAX": {
+        "wban": "3983",
+        "staid": "KEAX",
+        "stname": "Kansas City, MO",
+        "latitude": "38.81027778",
+        "longitude": "-94.26444444",
+        "elevation": "995",
+        "height": "20"
+    },
+    "KBYX": {
+        "wban": "92804",
+        "staid": "KBYX",
+        "stname": "Key West, FL",
+        "latitude": "24.5975",
+        "longitude": "-81.70305556",
+        "elevation": "8",
+        "height": "20"
+    },
+    "KMRX": {
+        "wban": "53832",
+        "staid": "KMRX",
+        "stname": "Knoxville/Tri-Cities, TN",
+        "latitude": "36.16861111",
+        "longitude": "-83.40166667",
+        "elevation": "1337",
+        "height": "20"
+    },
+    "KARX": {
+        "wban": "94987",
+        "staid": "KARX",
+        "stname": "La Crosse, WI",
+        "latitude": "43.82277778",
+        "longitude": "-91.19111111",
+        "elevation": "1276",
+        "height": "20"
+    },
+    "LPLA": {
+        "wban": "PENDING",
+        "staid": "LPLA",
+        "stname": "Lajes AB, Azores",
+        "latitude": "38.73027778",
+        "longitude": "-27.32166667",
+        "elevation": "3334",
+        "height": "20"
+    },
+    "KLCH": {
+        "wban": "3937",
+        "staid": "KLCH",
+        "stname": "Lake Charles, LA",
+        "latitude": "30.12527778",
+        "longitude": "-93.21583333",
+        "elevation": "13",
+        "height": "20"
+    },
+    "KESX": {
+        "wban": "53110",
+        "staid": "KESX",
+        "stname": "Las Vegas, NV",
+        "latitude": "35.70111111",
+        "longitude": "-114.8913889",
+        "elevation": "4867",
+        "height": "20"
+    },
+    "KDFX": {
+        "wban": "22015",
+        "staid": "KDFX",
+        "stname": "Laughlin AFB, TX",
+        "latitude": "29.27277778",
+        "longitude": "-100.2805556",
+        "elevation": "1131",
+        "height": "10"
+    },
+    "KILX": {
+        "wban": "4833",
+        "staid": "KILX",
+        "stname": "Lincoln, IL",
+        "latitude": "40.15055556",
+        "longitude": "-89.33694444",
+        "elevation": "582",
+        "height": "30"
+    },
+    "KLZK": {
+        "wban": "3952",
+        "staid": "KLZK",
+        "stname": "Little Rock, AR",
+        "latitude": "34.83638889",
+        "longitude": "-92.26222222",
+        "elevation": "568",
+        "height": "20"
+    },
+    "KVTX": {
+        "wban": "53101",
+        "staid": "KVTX",
+        "stname": "Los Angeles, CA",
+        "latitude": "34.41166667",
+        "longitude": "-119.1794444",
+        "elevation": "2726",
+        "height": "20"
+    },
+    "KLVX": {
+        "wban": "53827",
+        "staid": "KLVX",
+        "stname": "Louisville, KY",
+        "latitude": "37.97527778",
+        "longitude": "-85.94388889",
+        "elevation": "719",
+        "height": "30"
+    },
+    "KLBB": {
+        "wban": "23042",
+        "staid": "KLBB",
+        "stname": "Lubbock, TX",
+        "latitude": "33.65416667",
+        "longitude": "-101.8141667",
+        "elevation": "3259",
+        "height": "20"
+    },
+    "KMQT": {
+        "wban": "94850",
+        "staid": "KMQT",
+        "stname": "Marquette, MI",
+        "latitude": "46.53111111",
+        "longitude": "-87.54833333",
+        "elevation": "1411",
+        "height": "30"
+    },
+    "KMXX": {
+        "wban": "53826",
+        "staid": "KMXX",
+        "stname": "Maxwell AFB, AL",
+        "latitude": "32.53666667",
+        "longitude": "-85.78972222",
+        "elevation": "400",
+        "height": "30"
+    },
+    "KMAX": {
+        "wban": "94296",
+        "staid": "KMAX",
+        "stname": "Medford, OR",
+        "latitude": "42.08111111",
+        "longitude": "-122.7172222",
+        "elevation": "7513",
+        "height": "10"
+    },
+    "KMLB": {
+        "wban": "12838",
+        "staid": "KMLB",
+        "stname": "Melbourne, FL",
+        "latitude": "28.11333333",
+        "longitude": "-80.65416667",
+        "elevation": "35",
+        "height": "20"
+    },
+    "KNQA": {
+        "wban": "93839",
+        "staid": "KNQA",
+        "stname": "Memphis, TN",
+        "latitude": "35.34472222",
+        "longitude": "-89.87333333",
+        "elevation": "282",
+        "height": "20"
+    },
+    "KAMX": {
+        "wban": "12899",
+        "staid": "KAMX",
+        "stname": "Miami, FL",
+        "latitude": "25.61111111",
+        "longitude": "-80.41277778",
+        "elevation": "14",
+        "height": "20"
+    },
+    "KMAF": {
+        "wban": "23023",
+        "staid": "KMAF",
+        "stname": "Midland/Odessa, TX",
+        "latitude": "31.94333333",
+        "longitude": "-102.1891667",
+        "elevation": "2868",
+        "height": "10"
+    },
+    "KMKX": {
+        "wban": "4834",
+        "staid": "KMKX",
+        "stname": "Milwaukee, WI",
+        "latitude": "42.96777778",
+        "longitude": "-88.55055556",
+        "elevation": "958",
+        "height": "10"
+    },
+    "KMPX": {
+        "wban": "94983",
+        "staid": "KMPX",
+        "stname": "Minneapolis/St. Paul, MN",
+        "latitude": "44.84888889",
+        "longitude": "-93.56555556",
+        "elevation": "946",
+        "height": "30"
+    },
+    "KMBX": {
+        "wban": "94048",
+        "staid": "KMBX",
+        "stname": "Minot AFB, ND",
+        "latitude": "48.3925",
+        "longitude": "-100.865",
+        "elevation": "1493",
+        "height": "20"
+    },
+    "KMSX": {
+        "wban": "4103",
+        "staid": "KMSX",
+        "stname": "Missoula, MT",
+        "latitude": "47.04111111",
+        "longitude": "-113.9861111",
+        "elevation": "7855",
+        "height": "10"
+    },
+    "KMOB": {
+        "wban": "13894",
+        "staid": "KMOB",
+        "stname": "Mobile, AL",
+        "latitude": "30.67944444",
+        "longitude": "-88.23972222",
+        "elevation": "208",
+        "height": "20"
+    },
+    "PHMO": {
+        "wban": "22547",
+        "staid": "PHMO",
+        "stname": "Molokai, HI",
+        "latitude": "21.13277778",
+        "longitude": "-157.18",
+        "elevation": "1363",
+        "height": "20"
+    },
+    "KVAX": {
+        "wban": "53856",
+        "staid": "KVAX",
+        "stname": "Moody AFB, GA",
+        "latitude": "30.39027778",
+        "longitude": "-83.00166667",
+        "elevation": "173",
+        "height": "30"
+    },
+    "KMHX": {
+        "wban": "93768",
+        "staid": "KMHX",
+        "stname": "Morehead City, NC",
+        "latitude": "34.77611111",
+        "longitude": "-76.87611111",
+        "elevation": "31",
+        "height": "30"
+    },
+    "KOHX": {
+        "wban": "53833",
+        "staid": "KOHX",
+        "stname": "Nashville, TN",
+        "latitude": "36.24722222",
+        "longitude": "-86.5625",
+        "elevation": "579",
+        "height": "20"
+    },
+    "KLIX": {
+        "wban": "53813",
+        "staid": "KLIX",
+        "stname": "New Orleans, LA",
+        "latitude": "30.33666667",
+        "longitude": "-89.82555556",
+        "elevation": "24",
+        "height": "30"
+    },
+    "KOKX": {
+        "wban": "94703",
+        "staid": "KOKX",
+        "stname": "New York City, NY",
+        "latitude": "40.86555556",
+        "longitude": "-72.86388889",
+        "elevation": "85",
+        "height": "30"
+    },
+    "KAKQ": {
+        "wban": "93773",
+        "staid": "KAKQ",
+        "stname": "Norfolk/Richmond, VA",
+        "latitude": "36.98388889",
+        "longitude": "-77.00722222",
+        "elevation": "112",
+        "height": "20"
+    },
+    "KLNX": {
+        "wban": "94049",
+        "staid": "KLNX",
+        "stname": "North Platte, NE",
+        "latitude": "41.95777778",
+        "longitude": "-100.5763889",
+        "elevation": "2970",
+        "height": "20"
+    },
+    "KTLX": {
+        "wban": "3979",
+        "staid": "KTLX",
+        "stname": "Oklahoma City, OK",
+        "latitude": "35.33305556",
+        "longitude": "-97.27777778",
+        "elevation": "1213",
+        "height": "15"
+    },
+    "KOAX": {
+        "wban": "94980",
+        "staid": "KOAX",
+        "stname": "Omaha, NE",
+        "latitude": "41.32027778",
+        "longitude": "-96.36666667",
+        "elevation": "1148",
+        "height": "30"
+    },
+    "KPAH": {
+        "wban": "3816",
+        "staid": "KPAH",
+        "stname": "Paducah, KY",
+        "latitude": "37.06833333",
+        "longitude": "-88.77194444",
+        "elevation": "392",
+        "height": "30"
+    },
+    "KPDT": {
+        "wban": "24155",
+        "staid": "KPDT",
+        "stname": "Pendleton, OR",
+        "latitude": "45.69055556",
+        "longitude": "-118.8527778",
+        "elevation": "1515",
+        "height": "10"
+    },
+    "KDIX": {
+        "wban": "93771",
+        "staid": "KDIX",
+        "stname": "Philadelphia, PA",
+        "latitude": "39.94694444",
+        "longitude": "-74.41083333",
+        "elevation": "149",
+        "height": "20"
+    },
+    "KIWA": {
+        "wban": "23104",
+        "staid": "KIWA",
+        "stname": "Phoenix, AZ",
+        "latitude": "33.28916667",
+        "longitude": "-111.67",
+        "elevation": "1353",
+        "height": "10"
+    },
+    "KPBZ": {
+        "wban": "4832",
+        "staid": "KPBZ",
+        "stname": "Pittsburgh, PA",
+        "latitude": "40.53166667",
+        "longitude": "-80.21805556",
+        "elevation": "1185",
+        "height": "20"
+    },
+    "KSFX": {
+        "wban": "4107",
+        "staid": "KSFX",
+        "stname": "Pocatello/Idaho Falls, ID",
+        "latitude": "43.10583333",
+        "longitude": "-112.6861111",
+        "elevation": "4474",
+        "height": "10"
+    },
+    "KGYX": {
+        "wban": "54762",
+        "staid": "KGYX",
+        "stname": "Portland, ME",
+        "latitude": "43.89138889",
+        "longitude": "-70.25638889",
+        "elevation": "409",
+        "height": "10"
+    },
+    "KRTX": {
+        "wban": "94288",
+        "staid": "KRTX",
+        "stname": "Portland, OR",
+        "latitude": "45.71472222",
+        "longitude": "-122.9652778",
+        "elevation": "1572",
+        "height": "30"
+    },
+    "KPUX": {
+        "wban": "3021",
+        "staid": "KPUX",
+        "stname": "Pueblo, CO",
+        "latitude": "38.45944444",
+        "longitude": "-104.1813889",
+        "elevation": "5249",
+        "height": "10"
+    },
+    "KRAX": {
+        "wban": "93772",
+        "staid": "KRAX",
+        "stname": "Raleigh/Durham, NC",
+        "latitude": "35.66555556",
+        "longitude": "-78.48972222",
+        "elevation": "348",
+        "height": "30"
+    },
+    "KUDX": {
+        "wban": "94047",
+        "staid": "KUDX",
+        "stname": "Rapid City, SD",
+        "latitude": "44.125",
+        "longitude": "-102.8297222",
+        "elevation": "3016",
+        "height": "30"
+    },
+    "KRGX": {
+        "wban": "53104",
+        "staid": "KRGX",
+        "stname": "Reno, NV",
+        "latitude": "39.75416667",
+        "longitude": "-119.4622222",
+        "elevation": "8299",
+        "height": "20"
+    },
+    "KRIW": {
+        "wban": "24061",
+        "staid": "KRIW",
+        "stname": "Riverton, WY",
+        "latitude": "43.06611111",
+        "longitude": "-108.4772222",
+        "elevation": "5568",
+        "height": "10"
+    },
+    "KFCX": {
+        "wban": "53831",
+        "staid": "KFCX",
+        "stname": "Roanoke, VA",
+        "latitude": "37.02444444",
+        "longitude": "-80.27388889",
+        "elevation": "2868",
+        "height": "20"
+    },
+    "KJGX": {
+        "wban": "53824",
+        "staid": "KJGX",
+        "stname": "Robins AFB, GA",
+        "latitude": "32.67527778",
+        "longitude": "-83.35111111",
+        "elevation": "521",
+        "height": "20"
+    },
+    "KDAX": {
+        "wban": "93235",
+        "staid": "KDAX",
+        "stname": "Sacramento, CA",
+        "latitude": "38.50111111",
+        "longitude": "-121.6777778",
+        "elevation": "30",
+        "height": "30"
+    },
+    "KLSX": {
+        "wban": "3982",
+        "staid": "KLSX",
+        "stname": "Saint Louis, MO",
+        "latitude": "38.69888889",
+        "longitude": "-90.68277778",
+        "elevation": "608",
+        "height": "30"
+    },
+    "KMTX": {
+        "wban": "4104",
+        "staid": "KMTX",
+        "stname": "Salt Lake City, UT",
+        "latitude": "41.26277778",
+        "longitude": "-112.4477778",
+        "elevation": "6460",
+        "height": "30"
+    },
+    "KSJT": {
+        "wban": "23034",
+        "staid": "KSJT",
+        "stname": "San Angelo, TX",
+        "latitude": "31.37138889",
+        "longitude": "-100.4925",
+        "elevation": "1890",
+        "height": "30"
+    },
+    "KNKX": {
+        "wban": "53115",
+        "staid": "KNKX",
+        "stname": "San Diego, CA",
+        "latitude": "32.91888889",
+        "longitude": "-117.0419444",
+        "elevation": "955",
+        "height": "20"
+    },
+    "KMUX": {
+        "wban": "93236",
+        "staid": "KMUX",
+        "stname": "San Francisco, CA",
+        "latitude": "37.15527778",
+        "longitude": "-121.8983333",
+        "elevation": "3469",
+        "height": "20"
+    },
+    "KHNX": {
+        "wban": "53111",
+        "staid": "KHNX",
+        "stname": "San Joaquin Valley, CA",
+        "latitude": "36.31416667",
+        "longitude": "-119.6322222",
+        "elevation": "243",
+        "height": "20"
+    },
+    "TJUA": {
+        "wban": "PENDING",
+        "staid": "TJUA",
+        "stname": "San Juan, PR",
+        "latitude": "18.11555556",
+        "longitude": "-66.07805556",
+        "elevation": "2794",
+        "height": "30"
+    },
+    "KSOX": {
+        "wban": "53117",
+        "staid": "KSOX",
+        "stname": "Santa Ana Mountains, CA",
+        "latitude": "33.81777778",
+        "longitude": "-117.6358333",
+        "elevation": "3027",
+        "height": "15"
+    },
+    "KATX": {
+        "wban": "94287",
+        "staid": "KATX",
+        "stname": "Seattle/Tacoma, WA",
+        "latitude": "48.19444444",
+        "longitude": "-122.4958333",
+        "elevation": "494",
+        "height": "30"
+    },
+    "KSHV": {
+        "wban": "13957",
+        "staid": "KSHV",
+        "stname": "Shreveport, LA",
+        "latitude": "32.45083333",
+        "longitude": "-93.84138889",
+        "elevation": "273",
+        "height": "30"
+    },
+    "KFSD": {
+        "wban": "14944",
+        "staid": "KFSD",
+        "stname": "Sioux Falls, SD",
+        "latitude": "43.58777778",
+        "longitude": "-96.72944444",
+        "elevation": "1430",
+        "height": "10"
+    },
+    "PHKI": {
+        "wban": "22548",
+        "staid": "PHKI",
+        "stname": "South Kauai, HI",
+        "latitude": "21.89416667",
+        "longitude": "-159.5522222",
+        "elevation": "179",
+        "height": "30"
+    },
+    "PHWA": {
+        "wban": "PENDING",
+        "staid": "PHWA",
+        "stname": "South Shore, HI",
+        "latitude": "19.095",
+        "longitude": "-155.5688889",
+        "elevation": "1380",
+        "height": "20"
+    },
+    "KOTX": {
+        "wban": "4106",
+        "staid": "KOTX",
+        "stname": "Spokane, WA",
+        "latitude": "47.68027778",
+        "longitude": "-117.6266667",
+        "elevation": "2384",
+        "height": "10"
+    },
+    "KSGF": {
+        "wban": "13995",
+        "staid": "KSGF",
+        "stname": "Springfield, MO",
+        "latitude": "37.23527778",
+        "longitude": "-93.40055556",
+        "elevation": "1278",
+        "height": "20"
+    },
+    "KCCX": {
+        "wban": "54764",
+        "staid": "KCCX",
+        "stname": "State College, PA",
+        "latitude": "40.92305556",
+        "longitude": "-78.00361111",
+        "elevation": "2405",
+        "height": "20"
+    },
+    "KLWX": {
+        "wban": "93767",
+        "staid": "KLWX",
+        "stname": "Sterling, VA",
+        "latitude": "38.97527778",
+        "longitude": "-77.47777778",
+        "elevation": "272",
+        "height": "25"
+    },
+    "KTLH": {
+        "wban": "93805",
+        "staid": "KTLH",
+        "stname": "Tallahassee, FL",
+        "latitude": "30.3975",
+        "longitude": "-84.32888889",
+        "elevation": "63",
+        "height": "30"
+    },
+    "KTBW": {
+        "wban": "92801",
+        "staid": "KTBW",
+        "stname": "Tampa, FL",
+        "latitude": "27.70555556",
+        "longitude": "-82.40166667",
+        "elevation": "41",
+        "height": "20"
+    },
+    "KTWX": {
+        "wban": "3986",
+        "staid": "KTWX",
+        "stname": "Topeka, KS",
+        "latitude": "38.99694444",
+        "longitude": "-96.2325",
+        "elevation": "1367",
+        "height": "10"
+    },
+    "KEMX": {
+        "wban": "53112",
+        "staid": "KEMX",
+        "stname": "Tucson, AZ",
+        "latitude": "31.89361111",
+        "longitude": "-110.6302778",
+        "elevation": "5205",
+        "height": "30"
+    },
+    "KINX": {
+        "wban": "3984",
+        "staid": "KINX",
+        "stname": "Tulsa, OK",
+        "latitude": "36.175",
+        "longitude": "-95.56472222",
+        "elevation": "668",
+        "height": "20"
+    },
+    "KVNX": {
+        "wban": "3995",
+        "staid": "KVNX",
+        "stname": "Vance AFB, OK",
+        "latitude": "36.74083333",
+        "longitude": "-98.12777778",
+        "elevation": "1210",
+        "height": "10"
+    },
+    "KVBX": {
+        "wban": "93234",
+        "staid": "KVBX",
+        "stname": "Vandenberg AFB, CA",
+        "latitude": "34.83805556",
+        "longitude": "-120.3969444",
+        "elevation": "1223",
+        "height": "20"
+    },
+    "KICT": {
+        "wban": "3928",
+        "staid": "KICT",
+        "stname": "Wichita, KS",
+        "latitude": "37.65472222",
+        "longitude": "-97.44277778",
+        "elevation": "1335",
+        "height": "10"
+    },
+    "KLTX": {
+        "wban": "93774",
+        "staid": "KLTX",
+        "stname": "Wilmington, NC",
+        "latitude": "33.98944444",
+        "longitude": "-78.42888889",
+        "elevation": "64",
+        "height": "20"
+    },
+    "KYUX": {
+        "wban": "53116",
+        "staid": "KYUX",
+        "stname": "Yuma, AZ",
+        "latitude": "32.49527778",
+        "longitude": "-114.6566667",
+        "elevation": "174",
+        "height": "10"
+    }
+}
